Migrate SessionController into UsersController
